Fix: fix install errors for windows w/o powershell
Before, powershell is used for install EnergyPlus on Windows silently.
However, it seems that there is no powershell install on some
cran-winbuilder.

As EnergyPlus installer is packaged using NullSoft "/S" option can be
used to let the installer run in quite mode. This does not require
powershell to be installed any more.